Добавляет модульные файлы
